feat: add support for form submission on Enter key press
This commit adds the ability to submit the registration form by pressing the Enter key while focused on any of the input fields.

The following changes were made:

- Created a new  function that listens for the Enter key press event
- Added the  event handler to each input field and passed the  function as the event handler
- Inside , checked if the pressed key is the Enter key ()
- If the Enter key is pressed, called the  function to trigger form submission

With this change, users can now submit the registration form by either clicking the Register button or pressing the Enter key while focused on any of the input fields, providing a more convenient user experience.